requirements.txt
Flask
opencv-python
numpy
Pillow
mediapipe
